+++ dart/geocore/lib/geocore.dart
@@ -18,3 +18,4 @@ export 'meta_extent.dart';
 export 'parse_factory.dart';
 export 'parse_geojson.dart';
 export 'parse_wkt.dart';
+export 'proj4.dart';