+++ dart/geocore/lib/proj4.dart
@@ -1,0 +1,12 @@
+// Copyright (c) 2020-2021 Navibyte (https://navibyte.com). All rights reserved.
+// Use of this source code is governed by a “BSD-3-Clause”-style license that is
+// specified in the LICENSE file.
+//
+// Docs: https://github.com/navibyte/geospatial
+
+/// Projections provided by the external `proj4dart` library.
+///
+/// Usage: import `package:geocore/proj4.dart`
+library proj4;
+
+export 'src/proj/proj4.dart';