+++ dart/geocore/lib/src/geo/webmercator.dart
@@ -17,73 +17,78 @@ import 'geopoint.dart';
 // https://docs.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system
 // https://athene-forschung.unibw.de/doc/132233/132233.pdf (Some Principles of Web Mercator Maps and their Computation)
 
-/// Returns a function that projects geographic points to web mercator points.
+/// A projection adapter between WGS84 geographic and Web Mercator points.
 ///
-/// By default, target points of [R] are created using [factory].
+/// Use `forward` of the adapter to return a conversion function as:
+/// * source: `lon` and `lat` coordinates ("EPSG:4326" / WGS84)
+/// * target: `x` and `y` coordinates ("EPSG:3857" / Web Mercator)
 ///
-/// The returned conversion function expects source coordinates to be in the
-/// `CRS84` (that is 'WGS 84 longitude-latitude' or EPSG:4326) coordinate
-/// reference system. Source `lon` and `lat` coordinates are converted to target
-/// `x` and `y` coordinates (meters) of the Web Mercator (EPSG:3857) projection.
-///
-/// Other coordinates, if available in the source and if expected to target
-/// coordinates, are then copied (`elev` => `z` and `m` => `m`) without any
-/// changes.
-ProjectPoint<R> wgs84ToWebMercator<R extends CartesianPoint>(
-  PointFactory<R> factory,
-) =>
-    (Point source, {PointFactory<R>? to}) {
-      // source coordinates
-      final lon = source.x; // longitude is at x
-      final lat = source.y; // latitude is at y
-
-      // project (lon, lat) to (x, y)
-      final x = lon * 20037508.34 / 180.0;
-      final y0 = math.log(math.tan((90.0 + lat) * math.pi / 360.0)) /
-          (math.pi / 180.0);
-      final y = y0 * 20037508.34 / 180;
-
-      // return a projected point with optional z and m coordinates unchanged
-      return (to ?? factory).newWith(
-        x: x,
-        y: y,
-        z: source.z,
-        m: source.m,
-      );
-    };
-
-/// Returns a function that unprojects web mercator points to geographic points.
-///
-/// By default, target points of [R] are created using [factory].
-///
-/// The returned conversion function expects source coordinates to be in the
-/// Web Mercator (EPSG:3857) projection. Source `x` and `y` coordinates (meters)
-/// are converted to target `lon` and `lat` coordinates of the `CRS84` (that is
-/// 'WGS 84 longitude-latitude' or EPSG:4326) coordinate reference system.
+/// Use `inverse` of the adapter to return a conversion function as:
+/// * source: `x` and `y` coordinates ("EPSG:3857" / Web Mercator)
+/// * target: `lon` and `lat` coordinates ("EPSG:4326" / WGS84)
 ///
 /// Other coordinates, if available in the source and if expected to target
-/// coordinates, are then copied  (`z` => `elev` and `m` => `m`) without any
+/// coordinates, are just copied (`elev` <=> `z` and `m` <=> `m`) without any
 /// changes.
-ProjectPoint<R> webMercatorToWgs84<R extends GeoPoint>(
-  PointFactory<R> factory,
-) =>
-    (Point source, {PointFactory<R>? to}) {
-      // source coordinates
-      final x = source.x;
-      final y = source.y;
+const ProjectionAdapter<GeoPoint, CartesianPoint> wgs84ToWebMercator =
+    _Wgs84ToWebMercatorAdapter();
 
-      // unproject (x, y) to (lon, lat)
-      final lon = (x / 20037508.34) * 180.0;
-      final lat0 = (y / 20037508.34) * 180.0;
-      final lat = 180.0 /
-          math.pi *
-          (2 * math.atan(math.exp(lat0 * math.pi / 180.0)) - math.pi / 2);
+class _Wgs84ToWebMercatorAdapter
+    with ProjectionAdapter<GeoPoint, CartesianPoint> {
+  const _Wgs84ToWebMercatorAdapter();
 
-      // return an unprojected point with optional z and m coordinates unchanged
-      return (to ?? factory).newWith(
-        x: lon,
-        y: lat,
-        z: source.z,
-        m: source.m,
-      );
-    };
+  @override
+  String get fromCode => 'EPSG:4326';
+
+  @override
+  String get toCode => 'EPSG:3857';
+
+  @override
+  ProjectPoint<R> forward<R extends CartesianPoint>(
+    PointFactory<R> factory,
+  ) =>
+      (Point source, {PointFactory<R>? to}) {
+        // source coordinates
+        final lon = source.x; // longitude is at x
+        final lat = source.y; // latitude is at y
+
+        // project (lon, lat) to (x, y)
+        final x = lon * 20037508.34 / 180.0;
+        final y0 = math.log(math.tan((90.0 + lat) * math.pi / 360.0)) /
+            (math.pi / 180.0);
+        final y = y0 * 20037508.34 / 180;
+
+        // return a projected point with optional z and m coordinates unchanged
+        return (to ?? factory).newWith(
+          x: x,
+          y: y,
+          z: source.z,
+          m: source.m,
+        );
+      };
+
+  @override
+  ProjectPoint<R> inverse<R extends GeoPoint>(
+    PointFactory<R> factory,
+  ) =>
+      (Point source, {PointFactory<R>? to}) {
+        // source coordinates
+        final x = source.x;
+        final y = source.y;
+
+        // unproject (x, y) to (lon, lat)
+        final lon = (x / 20037508.34) * 180.0;
+        final lat0 = (y / 20037508.34) * 180.0;
+        final lat = 180.0 /
+            math.pi *
+            (2 * math.atan(math.exp(lat0 * math.pi / 180.0)) - math.pi / 2);
+
+        // return an unprojected point with optional z and m coords unchanged
+        return (to ?? factory).newWith(
+          x: lon,
+          y: lat,
+          z: source.z,
+          m: source.m,
+        );
+      };
+}