+++ dart/geocore/lib/src/proj/proj4.dart
@@ -1,0 +1,1 @@
+export 'proj4/proj4dart_adapter.dart';